+++ test/Test/Codec/CBOR/Cuddle/CDDL/Examples/Huddle.hs
@@ -21,13 +21,11 @@ module Test.Codec.CBOR.Cuddle.CDDL.Examples.Huddle (
   listTooShortExample,
   listSkippedRuleExample,
   listSkippedRuleNestedExample,
-<<<<<<< HEAD
   mapLeftoverKVExample,
   mapNoMatchingKeyExample,
   listZeroOrMoreExample,
   mapNestedValueExample,
-=======
->>>>>>> d212208 (Improve list leftover elements validation trace)
+  deeplyNestedRefExample,
 ) where
 
 import Codec.CBOR.Cuddle.CDDL (Name)
@@ -289,3 +287,21 @@ mapNestedValueExample =
                 ]
         , HIRule bar
         ]
+
+-- | A rule with deeply nested references competing against a rule that makes
+-- actual term progress. Used to test that stacked references don't incorrectly
+-- beat rules with real progress.
+deeplyNestedRefExample :: Huddle
+deeplyNestedRefExample =
+  collectFrom
+    [ HIRule $ "root" =:= arr [a VInt, opt $ a aRef, opt $ a listRule]
+    , HIRule aRef
+    , HIRule barRef
+    , HIRule bazRef
+    , HIRule listRule
+    ]
+  where
+    bazRef = "baz" =:= VText
+    barRef = "bar" =:= bazRef
+    aRef = "a_ref" =:= barRef
+    listRule = "list_rule" =:= arr [a VInt, a VInt]